backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0.0
pydantic-settings
appwrite
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
